+++ src/mesh/targetimagepreviewmesh.ts
@@ -8,12 +8,15 @@ import { THREE } from "../three";
  * @see https://docs.zap.works/universal-ar/web-libraries/threejs/image-tracking/
  */
 export class TargetImagePreviewMesh extends THREE.Mesh {
-  public constructor(
-    target: ImageTarget,
-    material: THREE.Material = new THREE.MeshBasicMaterial({
-      map: new THREE.TextureLoader().load(target.image!.src),
-    })
-  ) {
+  public constructor(target: ImageTarget, material?: THREE.Material) {
+    if (!material) {
+      const map = new THREE.TextureLoader().load(target.image!.src);
+      map.colorSpace = THREE.SRGBColorSpace;
+      // eslint-disable-next-line no-param-reassign
+      material = new THREE.MeshBasicMaterial({
+        map,
+      });
+    }
     super(new TargetImagePreviewBufferGeometry(target), material);
   }
 }